Add more source comments to the code generator
I'm experimenting with adding Mako support to coverage.py.  In order to
fully parse the generated Python files to correlate generated lines with
source lines, I've added more comment output to identify the lines that
aren't associated with any source line.

This is probably not all the places that need to be commented, but seems
to work well for the templates I'm using. If you have a better way to do
this, or a different way to accomplish the same thing, I'm completely
open to other ideas.